feat(delaunay): Cleanup and renaming
