+++ packages/visx-delaunay/src/components/Polygon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
 
-export type VoronoiPolygonProps = {
+export type PolygonProps = {
   /** Override render function which is provided polygon and generated path. */
   children?: ({ path, polygon }: { path: string; polygon: [number, number][] }) => React.ReactNode;
   /** className to apply to path element. */
@@ -15,7 +15,7 @@ export default function Polygon({
   className,
   children,
   ...restProps
-}: VoronoiPolygonProps & Omit<React.SVGProps<SVGPathElement>, keyof VoronoiPolygonProps>) {
+}: PolygonProps & Omit<React.SVGProps<SVGPathElement>, keyof PolygonProps>) {
   if (!polygon) return null;
   const path = `M${polygon.join('L')}Z`;
   if (children) return <>{children({ path, polygon })}</>;
