+++ packages/visx-demo/src/sandboxes/exampleToVisxDependencyLookup.ts
@@ -9,7 +9,7 @@ import barstackHorizontalPackageJson from './visx-barstack-horizontal/package.js
 import brushPackageJson from './visx-brush/package.json';
 import chordPackageJson from './visx-chord/package.json';
 import curvePackageJson from './visx-curve/package.json';
-import delaunayPackageJson from './visx-delaunay/package.json';
+import delaunayPackageJson from './visx-delaunay-triangulation/package.json';
 import delaunayVoronoiPackageJson from './visx-delaunay-voronoi/package.json';
 import dendrogramPackageJson from './visx-dendrogram/package.json';
 import dotsPackageJson from './visx-dots/package.json';
